Move git utils to the git app
