+++ src/olympia/addons/tests/test_tasks.py
@@ -11,8 +11,8 @@ from olympia.amo.storage_utils import copy_stored_file
 from olympia.amo.tests import (
     addon_factory, TestCase, version_factory)
 from olympia.files.utils import id_to_path
+from olympia.git.utils import AddonGitRepository
 from olympia.versions.models import VersionPreview
-from olympia.lib.git import AddonGitRepository
 
 
 class TestMigrateWebextensionsToGitStorage(TestCase):
